+++ spigot/src/test/java/net/apartium/cocoabeans/spigot/listener/AnotherListener.java
@@ -1,6 +1,8 @@
 package net.apartium.cocoabeans.spigot.listener;
 
+import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -11,6 +13,11 @@ public class AnotherListener implements Listener {
     public AnotherListener(JavaPlugin plugin) {
         if (plugin == null)
             throw new NullPointerException("plugin cannot be null");
+    }
+
+    @EventHandler
+    public void on(PlayerJoinEvent event) {
+
     }
 
 }