+++ spigot/src/test/java/net/apartium/cocoabeans/spigot/listener/FailedListener.java
@@ -1,6 +1,8 @@
 package net.apartium.cocoabeans.spigot.listener;
 
+import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerJoinEvent;
 
 import static org.junit.jupiter.api.Assertions.fail;
 
@@ -8,6 +10,11 @@ public class FailedListener implements Listener {
 
     public FailedListener() {
         throw new RuntimeException();
+    }
+
+    @EventHandler
+    public void on(PlayerJoinEvent event) {
+
     }
 
 }