+++ spigot/src/test/java/net/apartium/cocoabeans/spigot/listener/ListenerRegisterTest.java
@@ -4,10 +4,15 @@ import net.apartium.cocoabeans.CollectionHelpers;
 import net.apartium.cocoabeans.spigot.SpigotTestBase;
 import net.apartium.cocoabeans.spigot.lazies.ListenerAutoRegistration;
 import org.bukkit.Bukkit;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.HandlerList;
+import org.bukkit.event.Listener;
+import org.bukkit.plugin.RegisteredListener;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
 import java.util.logging.Handler;
@@ -30,39 +35,24 @@ public class ListenerRegisterTest extends SpigotTestBase {
 
         assertFalse(myManager.hasBeenCreated());
 
-        List<String> logs = new ArrayList();
-
-        Handler handler = new Handler() {
-
-
-            @Override
-            public void publish(LogRecord logRecord) {
-                if (!logRecord.getMessage().startsWith("Loaded listener "))
-                    return;
-
-                logs.add(logRecord.getMessage());
-            }
-
-            @Override
-            public void flush() {
-            }
-
-            @Override
-            public void close() throws SecurityException {
-
-            }
-        };
-
-        Bukkit.getLogger().addHandler(handler);
-
         listenerAutoRegistration.addInjectableObject(myManager);
 
         assertDoesNotThrow(() -> listenerAutoRegistration.register(getClass().getPackageName(), false, Set.of(FailedListener.class)));
 
-        Bukkit.getLogger().removeHandler(handler);
-
         assertTrue(myManager.hasBeenCreated());
-        assertTrue(CollectionHelpers.equalsList(List.of("Loaded listener MyListener!", "Loaded listener AnotherListener!", "Loaded listener JustListener!"), logs));
+
+        List<Class<? extends Listener>> list = HandlerList.getHandlerLists().stream()
+                .map(HandlerList::getRegisteredListeners)
+                .flatMap(Arrays::stream)
+                .map(RegisteredListener::getListener)
+                .<Class<? extends Listener>>map(Listener::getClass)
+                .toList();
+
+        assertFalse(list.contains(FailedListener.class));
+        assertTrue(list.contains(AnotherListener.class));
+        assertTrue(list.contains(JustListener.class));
+        assertTrue(list.contains(MyListener.class));
+
     }
 
     @Override